Entry Points

+++ src/app/beer_garden/app.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 import logging
+import multiprocessing
 from datetime import timedelta
 from functools import partial
 
 import brewtils.models
-import time
 from apscheduler.executors.pool import ThreadPoolExecutor as APThreadPoolExecutor
 from apscheduler.schedulers.background import BackgroundScheduler
 from brewtils.models import Events
@@ -15,8 +15,10 @@ from pytz import utc
 from requests.exceptions import RequestException
 
 import beer_garden
+import beer_garden.api
 import beer_garden.db.api as db
 import beer_garden.queue.api as queue
+from beer_garden.api.entry_point import EntryPoint
 from beer_garden.bg_utils.event_publisher import EventPublishers, EventPublisher
 from beer_garden.bg_utils.publishers import MongoPublisher
 from beer_garden.db.mongo.jobstore import MongoJobStore
@@ -24,52 +26,51 @@ from beer_garden.db.mongo.pruner import MongoPruner
 from beer_garden.local_plugins.loader import LocalPluginLoader
 from beer_garden.local_plugins.manager import LocalPluginsManager
 from beer_garden.local_plugins.monitor import LocalPluginMonitor
-from beer_garden.local_plugins.registry import LocalPluginRegistry
-from beer_garden.local_plugins.validator import LocalPluginValidator
-from beer_garden.log import load_plugin_log_config
+from beer_garden.log import load_plugin_log_config, EntryPointLogger
 from beer_garden.metrics import PrometheusServer
 from beer_garden.monitor import PluginStatusMonitor
-from beer_garden.requests import RequestValidator
 
 
-class BartenderApp(StoppableThread):
-    """Main Application that Runs the Beergarden Backend."""
+class Application(StoppableThread):
+    """Main Beer-garden application
+
+    This class is basically a wrapper around the various singletons that need to exist
+    in order for Beer-garden to function.
+
+    """
+
+    request_validator = None
+    scheduler = None
+    event_publishers = None
+    plugin_manager = None
+    clients = None
+    helper_threads = None
+    context = None
+    log_queue = None
+    log_reader = None
+    entry_points = []
 
     def __init__(self):
-        self.logger = logging.getLogger(__name__)
+        super(Application, self).__init__(
+            name="Application", logger=logging.getLogger(__name__)
+        )
 
-        self.request_validator = RequestValidator(beer_garden.config.get("validator"))
-        self.plugin_registry = LocalPluginRegistry()
-        self.plugin_validator = LocalPluginValidator()
+        self.initialize()
+
+    def initialize(self):
+        """Actually construct all the various component pieces"""
         self.scheduler = self._setup_scheduler()
         self.event_publishers = EventPublishers()
 
         load_plugin_log_config()
 
-        self.plugin_loader = LocalPluginLoader(
-            validator=self.plugin_validator,
-            registry=self.plugin_registry,
-            local_plugin_dir=beer_garden.config.get("plugin.local.directory"),
-            plugin_log_directory=beer_garden.config.get("plugin.local.log_directory"),
-            connection_info=beer_garden.config.get("entry.http"),
-            username=beer_garden.config.get("plugin.local.auth.username"),
-            password=beer_garden.config.get("plugin.local.auth.password"),
-        )
-
         self.plugin_manager = LocalPluginsManager(
-            loader=self.plugin_loader,
-            validator=self.plugin_validator,
-            registry=self.plugin_registry,
-            shutdown_timeout=beer_garden.config.get("plugin.local.timeout.shutdown"),
+            shutdown_timeout=beer_garden.config.get("plugin.local.timeout.shutdown")
         )
 
         plugin_config = beer_garden.config.get("plugin")
         self.helper_threads = [
-            HelperThread(
-                LocalPluginMonitor,
-                plugin_manager=self.plugin_manager,
-                registry=self.plugin_registry,
-            ),
+            HelperThread(LocalPluginMonitor, plugin_manager=self.plugin_manager),
             HelperThread(
                 PluginStatusMonitor,
                 timeout_seconds=plugin_config.status_timeout,
@@ -98,25 +99,36 @@ class BartenderApp(StoppableThread):
                 )
             )
 
-        super(BartenderApp, self).__init__(logger=self.logger, name="BartenderApp")
+        self.context = multiprocessing.get_context("spawn")
+        self.log_queue = self.context.Queue()
+        self.log_reader = HelperThread(EntryPointLogger, log_queue=self.log_queue)
+
+        for entry_name, entry_value in beer_garden.config.get("entry").items():
+            if entry_value.get("enable"):
+                self.entry_points.append(EntryPoint.create(entry_name))
 
     def run(self):
         self._startup()
 
-        while not self.stopped():
-            for helper_thread in self.helper_threads:
-                if not helper_thread.thread.is_alive():
-                    self.logger.warning(
-                        "%s is dead, restarting" % helper_thread.display_name
-                    )
-                    helper_thread.start()
-
-            time.sleep(0.1)
+        while not self.wait(0.1):
+            for helper in self.helper_threads:
+                if not helper.thread.is_alive():
+                    self.logger.warning(f"{helper.display_name} is dead, restarting")
+                    helper.start()
 
         self._shutdown()
 
+    def _progressive_backoff(self, func, failure_message):
+        wait_time = 0.1
+        while not self.stopped() and not func():
+            self.logger.warning(failure_message)
+            self.logger.warning("Waiting %.1f seconds before next attempt", wait_time)
+
+            self.wait(wait_time)
+            wait_time = min(wait_time * 2, 30)
+
     def _startup(self):
-        self.logger.info("Starting Bartender...")
+        self.logger.debug("Starting Application...")
 
         self.logger.debug("Setting up database...")
         self._setup_database()
@@ -124,21 +136,28 @@ class BartenderApp(StoppableThread):
         self.logger.debug("Setting up queues...")
         self._setup_queues()
 
-        self.logger.info("Starting event publishers")
+        self.logger.debug("Starting event publishers")
         self.event_publishers = self._setup_event_publishers()
         # self.event_publishers = self._setup_event_publishers(client_ssl)
 
-        self.logger.info("Starting helper threads...")
+        self.logger.debug("Starting helper threads...")
         for helper_thread in self.helper_threads:
             helper_thread.start()
 
-        self.logger.info("Loading all local plugins...")
-        self.plugin_loader.load_plugins()
+        self.logger.info("Starting log reader...")
+        self.log_reader.start()
 
-        self.logger.info("Starting all local plugins...")
+        self.logger.debug("Starting entry points...")
+        for entry_point in self.entry_points:
+            entry_point.start(context=self.context, log_queue=self.log_queue)
+
+        self.logger.debug("Loading all local plugins...")
+        LocalPluginLoader.instance().load_plugins()
+
+        self.logger.debug("Starting all local plugins...")
         self.plugin_manager.start_all_plugins()
 
-        self.logger.info("Starting scheduler")
+        self.logger.debug("Starting scheduler")
         self.scheduler.start()
 
         try:
@@ -149,23 +168,25 @@ class BartenderApp(StoppableThread):
         except RequestException:
             self.logger.warning("Unable to publish startup notification")
 
-        self.logger.info("Beer-garden startup completed")
+        self.logger.info("All set! Let me know if you need anything else!")
 
     def _shutdown(self):
-        self.logger.info("Beginning Beer-garden shutdown process")
+        self.logger.info(
+            "Closing time! You don't have to go home, but you can't stay here."
+        )
 
         if self.scheduler.running:
-            self.logger.info("Pausing scheduler - no more jobs will be run")
+            self.logger.debug("Pausing scheduler - no more jobs will be run")
             self.scheduler.pause()
 
         self.plugin_manager.stop_all_plugins()
 
-        self.logger.info("Stopping helper threads")
+        self.logger.debug("Stopping helper threads")
         for helper_thread in reversed(self.helper_threads):
             helper_thread.stop()
 
         if self.scheduler.running:
-            self.logger.info("Shutting down scheduler")
+            self.logger.debug("Shutting down scheduler")
             self.scheduler.shutdown(wait=False)
 
         try:
@@ -176,16 +197,50 @@ class BartenderApp(StoppableThread):
         except RequestException:
             self.logger.warning("Unable to publish shutdown notification")
 
+        self.logger.debug("Stopping entry points")
+        for entry_point in self.entry_points:
+            entry_point.stop(timeout=10)
+
+        self.logger.info("Stopping log reader")
+        self.log_reader.stop()
+
         self.logger.info("Successfully shut down Beer-garden")
 
-    @staticmethod
-    def _setup_database():
+    def _setup_database(self):
+        """Startup tasks related to the database
+
+        This will:
+          - ensure a connection to the database (using a connection with short timeouts)
+          - create the actual connection to the database
+          - ensure the database is in a good state
+        """
+        self._progressive_backoff(
+            partial(db.check_connection, beer_garden.config.get("db")),
+            "Unable to connect to mongo, is it started?",
+        )
         db.create_connection(db_config=beer_garden.config.get("db"))
         db.initial_setup(beer_garden.config.get("auth.guest_login_enabled"))
 
-    @staticmethod
-    def _setup_queues():
+    def _setup_queues(self):
+        """Startup tasks related to the message queues
+
+        This will:
+          - create the message queue clients
+          - make sure that all clients have active connections
+          - ensure the broker and queues are in a good state
+        """
         queue.create_clients(beer_garden.config.get("amq"))
+
+        self._progressive_backoff(
+            partial(queue.check_connection, "pika"),
+            "Unable to connect to rabbitmq, is it started?",
+        )
+        self._progressive_backoff(
+            partial(queue.check_connection, "pyrabbit"),
+            "Unable to connect to rabbitmq admin interface. "
+            "Is the management plugin enabled?",
+        )
+
         queue.initial_setup()
 
     @staticmethod
@@ -275,6 +330,10 @@ class HelperThread(object):
         self.thread.start()
 
     def stop(self):
+        # Thread was never started - nothing to do
+        if not getattr(self, "thread"):
+            return
+
         if not self.thread.is_alive():
             self.logger.warning(
                 "Uh-oh. Looks like a bad shutdown - the %s " "was already stopped",
